backend/requirements.txt
django-allauth==0.44.0
Django==3.2.2
django-multiselectfield==0.1.12
Pillow==8.0.1
python-dotenv
git+https://github.com/sapristi/django-simple-blog.git
meilisearch==0.14.2
mistune==2.0.0rc1
django_filter
django-cors-headers
more-itertools
psycopg2
fastapi
uvicorn
django-extensions
django-simple-bulma
praw
fuzzywuzzy
python-Levenshtein
cysystemd
